relayer: Return l1 transaction hash in isCompleted check of Arbitrum relayer service
